requirements.txt
python-bitcoinlib
nwdiag
redis
